DF-10338 Fix review issues
Former-commit-id: 470d62600c64d994497debc1ca6b024376de92fd

+++ Sources/Internal/Utils/TextBox.h
@@ -133,7 +133,7 @@ inline TextBox::Line& TextBox::GetLine(const uint32 index)
 
 inline uint32 TextBox::GetLinesCount() const
 {
-    return lines.size();
+    return uint32(lines.size());
 }
 
 inline const Vector<TextBox::Character>& TextBox::GetCharacters() const
@@ -158,6 +158,6 @@ inline TextBox::Character& TextBox::GetCharacter(const uint32 index)
 
 inline uint32 TextBox::GetCharactersCount() const
 {
-    return characters.size();
+    return uint32(characters.size());
 }
 }
